pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<3.15"
aiogram = "^3.26.0"
python-dotenv = "^1.2.2"
aiohttp-socks = "^0.11.0"
httpx = "^0.28.1"
beautifulsoup4 = "^4.14.3"
lxml = "^6.0.2"
sqlalchemy = "^2.0.48"
psycopg2 = "^2.9.11"
alembic = "^1.18.4"
asyncpg = "^0.31.0"
pytz = "^2026.1.post1"
singbox2proxy = "^0.2.7"
ruff = "^0.15.8"
mypy = "^1.19.1"
redis = "^7.4.0"
pytest = "^9.0.3"
pytest-asyncio = "^1.3.0"
pytest-mock = "^3.15.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py314"
exclude = ["venv", "versions"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = 3.14
ignore_missing_imports = true
disallow_untyped_defs = true
strict = true
exclude = [
    '^tests/',
]

[[tool.mypy.overrides]]
module = "src.bot.database.orm_models"
ignore_errors = true